Refactor ReportLayerPanel
