+++ applications/fishing-map/src/features/workspace/shared/DatasetSchemaField.tsx
@@ -1,0 +1,34 @@
+import React, { Fragment } from 'react'
+import { TagList } from '@globalfishingwatch/ui-components'
+import { UrlDataviewInstance } from 'types'
+import styles from 'features/workspace/shared/LayerPanel.module.css'
+import {
+  getSchemaFieldsSelectedInDataview,
+  SupportedDatasetSchema,
+} from 'features/datasets/datasets.utils'
+
+type LayerPanelProps = {
+  dataview: UrlDataviewInstance
+  field: SupportedDatasetSchema
+  label: string
+}
+
+function DatasetSchemaField({ dataview, field, label }: LayerPanelProps): React.ReactElement {
+  const valuesSelected = getSchemaFieldsSelectedInDataview(dataview, field)
+  return (
+    <Fragment>
+      {valuesSelected.length > 0 && (
+        <div className={styles.filter}>
+          <label>{label}</label>
+          <TagList
+            tags={valuesSelected}
+            color={dataview.config?.color}
+            className={styles.tagList}
+          />
+        </div>
+      )}
+    </Fragment>
+  )
+}
+
+export default DatasetSchemaField
